Created "components" folder and moved all components there. Added very simple nav bar component
Created "components" folder and moved all components there. Added very simple nav bar component.
