+++ rpi-eventhub/src/components/App/App.test.js
